Moved static StructsAndEnums to csharp file
