+++ scripts/generate-cocoa-bindings.ps1
@@ -117,94 +117,6 @@ $Text = $Text -replace '\bpublic\b', 'internal'
 # Remove static CFunctions class
 $Text = $Text -replace '(?ms)\nstatic class CFunctions.*?}\n', ''
 
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryFeedbackSource = @'
-[Native]
-internal enum SentryFeedbackSource : long
-{
-    Unknown = 0,
-    User = 1,
-    System = 2,
-    Other = 3
-}
-'@
-
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryLevel = @'
-
-[Native]
-internal enum SentryLevel : ulong
-{
-    None = 0,
-    Debug = 1,
-    Info = 2,
-    Warning = 3,
-    Error = 4,
-    Fatal = 5
-}
-'@
-
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryTransactionNameSource = @'
-
-[Native]
-internal enum SentryTransactionNameSource : long
-{
-    Custom = 0,
-    Url = 1,
-    Route = 2,
-    View = 3,
-    Component = 4,
-    Task = 5
-}
-'@
-
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryReplayQuality = @'
-[Native]
-internal enum SentryReplayQuality : long
-{
-    Low = 0,
-    Medium = 1,
-    High = 2
-}
-'@
-
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryReplayType = @'
-[Native]
-internal enum SentryReplayType : long
-{
-    Session = 0,
-    Buffer = 1
-}
-'@
-
-# This enum resides in the Sentry-Swift.h
-# Appending it here so we don't need to import and create bindings for the entire header
-$SentryRRWebEventType = @'
-[Native]
-internal enum SentryRRWebEventType : long
-{
-    None = 0,
-    Touch = 3,
-    Meta = 4,
-    Custom = 5
-}
-'@
-
-$Text += "`n$SentryFeedbackSource"
-$Text += "`n$SentryLevel"
-$Text += "`n$SentryTransactionNameSource"
-$Text += "`n$SentryReplayQuality"
-$Text += "`n$SentryReplayType"
-$Text += "`n$SentryRRWebEventType"
-
 # Add header and output file
 $Text = "$Header`n`n$Text"
 $Text | Out-File "$BindingsPath/$File"
