+++ src/Sentry.Bindings.Cocoa/Sentry.Bindings.Cocoa.csproj
@@ -33,6 +33,7 @@
     <ObjcBindingApiDefinition Include="PrivateApiDefinitions.cs" />
     <ObjcBindingApiDefinition Include="SwiftApiDefinitions.cs" />
     <ObjcBindingCoreSource Include="StructsAndEnums.cs" />
+    <ObjcBindingCoreSource Include="SwiftStructsAndEnums.cs" />
     <NativeReference Include="$(SentryCocoaFramework)" Kind="Framework" />
 
     <!-- Use a separate readme file in the nuget. -->
